refactor: changed directory disposition
