+++ day00/correction/03_populate.py
