+++ day00/correction/06_k_first.py
