+++ day00/correction/07_seniors.py
