+++ day00/correction/08_battle_royale.py
