+++ day00/correction/10_sweet_spot.py
