+++ day00/correction/12_worldwide.py
